requirements.txt
streamlit
crewai
crewai-tools
plotly
pandas
python-dotenv
litellm
groq
